+++ packages/fonts/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.3.12 (2024-02-13)
+
+**Note:** Version bump only for package @un/fonts
+
+
+
+
+
+## 1.3.11 (2024-02-13)
+
+**Note:** Version bump only for package @un/fonts
+
+
+
+
+
+## 1.3.10 (2024-01-26)
+
+**Note:** Version bump only for package @un/fonts
+
+
+
+
+
 ## 1.3.9 (2024-01-23)
 
 **Note:** Version bump only for package @un/fonts