+++ packages/icons-react/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 10.37.10 (2024-02-13)
+
+**Note:** Version bump only for package @wfp/icons-react
+
+
+
+
+
+## 10.37.9 (2024-02-13)
+
+**Note:** Version bump only for package @wfp/icons-react
+
+
+
+
+
+## 10.37.8 (2024-01-26)
+
+**Note:** Version bump only for package @wfp/icons-react
+
+
+
+
+
 ## 10.37.7 (2024-01-23)
 
 **Note:** Version bump only for package @wfp/icons-react